Move hook to hooks directory.
